Allow a UWP CCGameView to be be added from VS designer toolbox and embedded into the MainPage Grid.
Modify UWP Test application to use CCGameView from MainPage instead of creating manually and setting the game view.  XAML designer now works.
