+++ tests/tests/WindowsUAP/App.xaml.cs
@@ -45,10 +45,7 @@ namespace CocosSharp.Tests.WindowsUWP
         /// <param name="e">Details about the launch request and process.</param>
         protected override void OnLaunched(LaunchActivatedEventArgs e)
         {
-            var gamePage = new CCGameView();
-            gamePage.ViewCreated += LoadGame;
 
-            AppDelegate.SharedWindow = gamePage;
 #if DEBUG
             if (System.Diagnostics.Debugger.IsAttached)
             {
@@ -56,53 +53,36 @@ namespace CocosSharp.Tests.WindowsUWP
             }
 #endif
 
-            //Frame rootFrame = Window.Current.Content as Frame;
+            Frame rootFrame = Window.Current.Content as Frame;
 
-            //// Do not repeat app initialization when the Window already has content,
-            //// just ensure that the window is active
-            //if (rootFrame == null)
-            //{
-            //    // Create a Frame to act as the navigation context and navigate to the first page
-            //    rootFrame = new Frame();
+            // Do not repeat app initialization when the Window already has content,
+            // just ensure that the window is active
+            if (rootFrame == null)
+            {
+                // Create a Frame to act as the navigation context and navigate to the first page
+                rootFrame = new Frame();
 
-            //    rootFrame.NavigationFailed += OnNavigationFailed;
+                rootFrame.NavigationFailed += OnNavigationFailed;
 
-            //    if (e.PreviousExecutionState == ApplicationExecutionState.Terminated)
-            //    {
-            //        //TODO: Load state from previously suspended application
-            //    }
+                if (e.PreviousExecutionState == ApplicationExecutionState.Terminated)
+                {
+                    //TODO: Load state from previously suspended application
+                }
 
-            //    // Place the frame in the current Window
-            //    Window.Current.Content = rootFrame;
-            //}
+                // Place the frame in the current Window
+                Window.Current.Content = rootFrame;
+            }
 
-            //if (rootFrame.Content == null)
-            //{
-            //    // When the navigation stack isn't restored navigate to the first page,
-            //    // configuring the new page by passing required information as a navigation
-            //    // parameter
-            //    rootFrame.Navigate(typeof(MainPage), e.Arguments);
-            //}
-
-            Window.Current.Content = gamePage;
+            if (rootFrame.Content == null)
+            {
+                // When the navigation stack isn't restored navigate to the first page,
+                // configuring the new page by passing required information as a navigation
+                // parameter
+                rootFrame.Navigate(typeof(MainPage), e.Arguments);
+            }
 
             // Ensure the current window is active
             Window.Current.Activate();
-        }
-
-        void LoadGame(object sender, System.EventArgs e)
-        {
-            CCGameView gameView = sender as CCGameView;
-
-            if (gameView != null)
-            {
-                gameView.DesignResolution = new CCSizeI(1024, 768);
-                gameView.Stats.Enabled = true;
-                gameView.ResolutionPolicy = CCViewResolutionPolicy.ShowAll;
-                CCScene gameScene = new CCScene(gameView);
-                gameScene.AddLayer(new TestController());
-                gameView.RunWithScene(gameScene);
-            }
         }
 
         /// <summary>
