+++ tests/tests/WindowsUAP/MainPage.xaml
@@ -1,13 +1,14 @@
 ﻿<Page
-    x:Class="CocosSharp.Tests.WindowsUWP.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:CocosSharp.Tests.WindowsUWP"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:CocosSharp="using:CocosSharp"
+    x:Class="CocosSharp.Tests.WindowsUWP.MainPage"
     mc:Ignorable="d">
 
     <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
+        <CocosSharp:CCGameView x:Name="GameView" Margin="0"/>
     </Grid>
 </Page>
